+++ package.json
@@ -9,7 +9,8 @@
     "rename": "node ./src/fs/rename.js",
     "delete": "node ./src/fs/delete.js",
     "list": "node ./src/fs/list.js",
-    "read": "node ./src/fs/read.js"
+    "read": "node ./src/fs/read.js",
+    "modules": "node ./src/modules/cjsToEsm.mjs"
   },
   "repository": {
     "type": "git",
