+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,34 @@
+import path, { dirname } from "path";
+import { fileURLToPath } from "url";
+import { release, version } from "os";
+import { createServer as createServerHttp } from "http";
+import { createRequire } from "module";
+import "./files/c.js";
+
+const require = createRequire(import.meta.url);
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = require("./files/a.json");
+} else {
+  unknownObject = require("./files/b.json");
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+  res.end("Request accepted");
+});
+
+export { unknownObject, createMyServer };
